Added rx error helpers to the core
