+++ app/src/androidTest/java/com/xmartlabs/template/common/SleepingAction.java
