+++ core/src/main/java/com/xmartlabs/base/core/rx/error/CompletableObserverWithErrorHandling.java
@@ -1,4 +1,4 @@
-package com.xmartlabs.template.common.rx;
+package com.xmartlabs.base.core.rx.error;
 
 import io.reactivex.CompletableObserver;
 import io.reactivex.disposables.Disposable;
