+++ core/src/main/java/com/xmartlabs/base/core/rx/error/MaybeObserverWithErrorHandling.java
@@ -1,28 +1,28 @@
-package com.xmartlabs.template.common.rx;
+package com.xmartlabs.base.core.rx.error;
 
-import io.reactivex.Observer;
+import io.reactivex.MaybeObserver;
 import io.reactivex.disposables.Disposable;
 import io.reactivex.functions.Consumer;
 import lombok.RequiredArgsConstructor;
 import timber.log.Timber;
 
 /**
- * Implementation of {@link Observer} that calls a {@link Consumer} function {@code onError}.
- * To be used as an Observer hook with RxJavaPlugins.
+ * Implementation of {@link MaybeObserver} that calls a {@link Consumer} function {@code onError}.
+ * To be used as a MaybeSubscribe hook with RxJavaPlugins.
  */
 @RequiredArgsConstructor
-public final class ObserverWithErrorHandling<T> implements Observer<T> {
-  private final Observer<T> observer;
+public final class MaybeObserverWithErrorHandling<T> implements MaybeObserver<T> {
+  private final MaybeObserver<T> maybeObserver;
   private final Consumer<? super Throwable> onErrorCallback;
 
   @Override
   public void onSubscribe(Disposable d) {
-    observer.onSubscribe(d);
+    maybeObserver.onSubscribe(d);
   }
 
   @Override
-  public void onNext(T t) {
-    observer.onNext(t);
+  public void onSuccess(T t) {
+    maybeObserver.onSuccess(t);
   }
 
   @Override
@@ -32,11 +32,11 @@ public final class ObserverWithErrorHandling<T> implements Observer<T> {
     } catch (Exception exception) {
       Timber.e(exception);
     }
-    observer.onError(e);
+    maybeObserver.onError(e);
   }
 
   @Override
   public void onComplete() {
-    observer.onComplete();
+    maybeObserver.onComplete();
   }
 }
